# tools/kotohiro.openapi.yaml
openapi: 3.0.1
info:
  title: kotohiro
  description: ''
  version: 1.0.0
...
paths:
...
  /talksessions/{talkSessionID}/opinions:
    post:
      summary: セッションに対して意見投稿
      deprecated: false
      description: parentOpinionIDがなければルートの意見として投稿される
      operationId: postOpinionPost
      tags:
        - opinion
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                parentOpinionID:
                  type: string
                  example: ''
                  nullable: true
                title:
                  type: string
                  example: ''
                  nullable: true
                opinionContent:
                  type: string
                  maxLength: 140
                  example: ''
                referenceURL:
                  type: string
                  format: url
                  example: ''
                  nullable: true
                picture:
                  type: string
                  format: binary
                  description: 参考画像。4MiBまで
                  example: ''
              required:
                - opinionContent
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
      security:
        - apikey-header-SessionId: []
      x-ogen-operation-group: Opinion
    get:
      summary: セッションに対する意見一覧
      deprecated: false
      description: ''
      operationId: getOpinionsForTalkSession
      tags:
        - opinion
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: sort
          in: query
          description: ''
          required: false
          schema:
            type: string
            enum:
              - latest
              - mostReply
              - oldest
            nullable: true
        - name: limit
          in: query
          description: ''
          required: false
          example: 10
          schema:
            type: integer
            nullable: true
        - name: offset
          in: query
          description: ''
          required: false
          example: 0
          schema:
            type: integer
            nullable: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  opinions:
                    type: array
                    items:
                      type: object
                      properties:
                        opinion: *ref_0
                        user: *ref_1
                        replyCount:
                          type: integer
                        myVoteType: &ref_4
                          $ref: '#/components/schemas/voteType'
                          description: 意見投稿主の意見。ルート意見の場合はここには何も入らない
                      required:
                        - opinion
                        - user
                        - replyCount
                        - myVoteType
                  pagination:
                    type: object
                    properties:
                      totalCount:
                        type: integer
                    required:
                      - totalCount
                required:
                  - opinions
                  - pagination
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: Opinion
...
  /talksessions/{talkSessionId}/analysis:
    get:
      summary: 分析結果一覧
      deprecated: false
      description: ''
      operationId: talkSessionAnalysis
      tags:
        - talk_session
      parameters:
        - name: talkSessionId
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  myPosition: &ref_3
                    $ref: '#/components/schemas/userGroupPosition'
                  positions:
                    type: array
                    items: *ref_3
                  groupOpinions:
                    type: array
                    items:
                      type: object
                      properties:
                        groupName:
                          type: string
                        groupId:
                          type: integer
                        opinions:
                          type: array
                          items:
                            type: object
                            properties:
                              opinion: *ref_0
                              user: *ref_1
                              agreeCount:
                                type: integer
                              disagreeCount:
                                type: integer
                              passCount:
                                type: integer
                            required:
                              - opinion
                              - user
                              - agreeCount
                              - disagreeCount
                              - passCount
                      required:
                        - groupName
                        - opinions
                        - groupId
                required:
                  - positions
                  - groupOpinions
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - code
                  - message
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: TalkSession
...
  /talksessions/{talkSessionId}/report:
    get:
      summary: セッションレポートを返す
      deprecated: false
      description: ''
      operationId: getTalkSessionReport
      tags:
        - talk_session
      parameters:
        - name: talkSessionId
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    type: string
                required:
                  - report
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: TalkSession
...
  /talksessions/{talkSessionID}/conclusion:
    get:
      summary: 結論取得
      deprecated: false
      description: ''
      operationId: getConclusion
      tags:
        - talk_session
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema: &ref_5
                $ref: '#/components/schemas/conclusion'
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: TalkSession
    post:
      summary: 結論投稿
      deprecated: false
      description: |-
        結論（conclusion）はセッションが終了した後にセッっションの作成者が投稿できる文章。
        セッションの流れやグループの分かれ方などに対するセッション作成者の感想やそれらの意見を受け、これからの方向性などを記入する。
      operationId: postConclusion
      tags:
        - talk_session
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 2
                  description: 結論本文。2文字以上
                  example: よかった！
              required:
                - content
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema: *ref_5
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
      security:
        - apikey-header-SessionId: []
      x-ogen-operation-group: TalkSession
  /talksessions/{talkSessionID}/timelines:
    get:
      summary: タイムライン取得
      deprecated: false
      description: タイムラインはセッション終了後にセッション作成者が設定できるその後の予定を知らせるもの
      operationId: getTimeLine
      tags:
        - timeline
      parameters:
        - name: talkSessionID
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: *ref_6
                required:
                  - items
          headers: {}
        '400':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
        '500':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: internal
                  value:
                    code: GENERAL-000
                    message: INTERNAL SERVER ERROR.
          headers: {}
      security: []
      x-ogen-operation-group: Timeline
...
components:
  schemas:
...
    voteType:
      type: string
      description: 意見の3種のステータス
      enum:
        - agree
        - disagree
        - pass
      default: unvote
...
    userGroupPosition:
      type: object
      properties:
        posX:
          type: number
        posY:
          type: number
        displayId:
          type: string
        displayName:
          type: string
        iconURL:
          type: string
          nullable: true
        groupName:
          type: string
        groupId:
          type: integer
        perimeterIndex:
          type: integer
          description: 境界ポイントのインデックス
      required:
        - posX
        - posY
        - displayId
        - groupName
        - displayName
        - groupId
...
    conclusion:
      type: object
      properties:
        user: *ref_1
        content:
          type: string
          description: 結論本文
      required:
        - user
        - content
...
  securitySchemes:
    apikey-header-SessionId:
      type: apiKey
      in: header
      name: SessionId
servers: []